fix(github): fix the plugin identification
This will write in a known bug-free plugin identification because the
previous implementation was a little bit messy and untested.

Each time that a language was found will create a new plugin,
this is wrong because it is just a guess that we can do this way,
and we can risk trying to be smart but we are really dumb like
in the case of javascript that it can contains also the typescript
configuration where in this case, we do not know what happens.

The work is not finished because the configuration should override
the guess about the language, because the conf information
is more important for the plugin manager.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>