feat(lib): add support for binary configuration
The tool for core lightning are not all plugin, but they can
be just binary like the coffee one.

So, this will start to support inside the configuration, the possibility
to support binaries installation and tracking.

So, a repository can have also custom binaries, like coffee or core
lightning itself.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>