fix(git): fix derivation of the exec path
we was forcing ourself to find the correct exec path, and we
was wrong! We can not always find the exec path, and when we are
not able to find it, we skip the plugin and we do not index it.

Simple enough for now, then later we should also think a little bit
on what happens if we upgrade coffee and we are able to index a plugin
that was not possible to index before.

I think the `coffee upgrade --refresh` should do the job?

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>