feat(lib): The repository can now upgrade itself
This commit provides a small improvement for @tareknaser360
pull request at https://github.com/coffee-tools/coffee/pull/132. Although
I'm not a fan of Rust due to its association with functional programming,
I have to admit that it is incredibly useful in terms of finding the
right solution.

Specifically, the commit at https://github.com/coffee-tools/coffee/pull/132/commits/3b57cae22b88c8ceef6138e4ead0ac54c030a7cf
does not make efficient use of the separation of concerns. In practice,
Coffee needs to request the repository to upgrade itself, and the repository
returns a list of plugins affected by this upgrade. Coffee then iterates over
these plugins and performs the installation procedure again.

This rework eliminates the need for the repository to be mutable (although it
will still modify the state of the disk, of course) and removes the plugin
after the upgrade is called.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>