fix(core): Create the network directory only when the user asks for it
This commit fixes a nasty bug that occurs when we use a network
that is not supported by coffee.

This bug was introduced in the proof-of-concept (PoC) of the plugin
manager, where we created a directory for each network.

With this commit, we create the directory lazily,
only when the user requires it.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>