fix(test): use the tmpdir in the correct way
This commit revert my previous solution to clone the tmpdir,
because it is wrong, the tmpdir wil be deleted when the first
instance is dropped.

So this commit use the Arc to pass around the tmp dir, so in this
way we keep a smart pointer and we solve all our problem.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>