fix(test): run core lightning with the plugin directly
Inside the CI we noted that the tests over the plugin get
stuck and my suspection is that we had some problem to run
the plugin with the `lightning-cli plugin start`.

So this is removing the method to start it when cln
will start, and running it with the deamon itself.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>