fix(core): store the repository globally
While working with coffee, we noted that there was a
limitation on how we store the repositories.

In fact, we store them once per network, and this causes problem
when we try to work with multiple networks,
and installing the same plugins.

Now with this commit, we store the repositories globally,
outside of the configuration network, so in this way we
can download once the remote is on and use
it in multiple directories.

Link: https://github.com/coffee-tools/coffee/issues/172
Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>