git: remove the libgit2 upgrade commands
This commit introduce a basic refactoring that remove the libgit2
solution to upgrade a repository and use the pure git command.

Currently we are using git in other places, and it should be fine to use
it there.

To keep faith in our contribution guide line, we should implement the
simple things that works, without implementing a complex things that did
not work.

At some point when coffee is stable we can thing to remove the git
command usage and use a pure git library.

Changelog-Fixed: git: remove the libgit2 upgrade commands and use a raw
git command
Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>