feat(coffee): add field branch to coffee install
This commit adds an optional parameter to coffee_install() method that makes it possible to specify a branch when installing a plugin.

The logic is:
- get information about current status before switching (repoisitory branch and plugin git head)
- switch to the branch
- restore old information and rollback

A function git_upgrade_with_git_head was implemented to handle checking out using a specfic git_head

we are using the repository git head to save current status and roll back to it since it's safer than using just the branch (we fetch origin so maybe the branch was updated. we want to be back to the same commit)

also, cmd, coffee_plugin and coffee httpd were upgraded to pass the branch value

Signed-off-by: Tarek <tareknaser360@gmail.com>